Added and Exceptions namespace
